[e2e] Move test support files into their own module
