+++ e2e/src/support/mod.rs
@@ -1,0 +1,9 @@
+// Copyright lowRISC contributors.
+// Licensed under the Apache License, Version 2.0, see LICENSE for details.
+// SPDX-License-Identifier: Apache-2.0
+
+//! Support utilities for Manticore e2e tests.
+
+pub mod fakes;
+pub mod rot;
+pub mod tcp;
